Fix __repr__ implementations in the ast wrapper
